workaround a weird chrome bug
when an input with type=password was in the page, the page flickered
each time on page load.  This commit work around that problem by
setting the input type to 'text' and by changing it back to 'password'
with some javascript.

It might be a problem with autocomplete, but that need some testing